Began gutting the blocking gameloop and replacing it with a poor man's event handling setup
