+++ Display.cs
@@ -1,15 +1,20 @@
-﻿using System.Runtime.InteropServices;
-
-using SDL2;
-using static SDL2.SDL;
-
-using BearLibNET;
-using BearLibNET.DefaultImplementations;
-using TKCodes = BearLibNET.TKCodes;
+﻿using BearLibNET.DefaultImplementations;
 
 namespace Yarl2;
 
-internal abstract class Display
+enum LoopState
+{
+    Startup,
+    Pregame,
+    Playing,
+    Quitting    
+}
+
+enum UIEventType { Quiting, KeyInput, NoEvent }
+
+internal record struct UIEvent(UIEventType Type, char Value);
+
+internal abstract class UserInterface
 {
     protected const int BACKSPACE = 8;
     protected const int ScreenWidth = 60;
@@ -42,18 +47,22 @@ internal abstract class Display
     public abstract char WaitForInput();
     public abstract void WriteLongMessage(List<string> message);
     public abstract void WriteMessage(string message);
+    protected abstract UIEvent PollForEvent();
+    
+    protected List<string>? _longMessage;
+    protected string _lastMessage = "";
     
     public Player? Player { get; set; } = null;
 
-    public Display()
+    public UserInterface()
     {
         PlayerScreenRow = (ScreenHeight - 1) / 2 + 1;
-        PlayerScreenCol = (ScreenWidth - SideBarWidth - 1) / 2;
+        PlayerScreenCol = (ScreenWidth - SideBarWidth - 1) / 2;        
     }
 
     public virtual void TitleScreen()
     {
-        var msg = new List<string>()
+        _longMessage = new List<string>()
         {
             "",
             "",
@@ -62,8 +71,7 @@ internal abstract class Display
             "     Welcome to Yarl2",
             "       (yet another attempt to make a roguelike,",
             "           this time in C#...)"
-        };
-        WriteLongMessage(msg);           
+        };        
     }
 
     private (int, int) AskForDirection()
@@ -185,444 +193,180 @@ internal abstract class Display
                 return (BLACK, ' ');
         }        
     }
-}
 
-internal class SDLDisplay : Display
-{
-    private readonly IntPtr _window;
-    private readonly IntPtr _renderer, _font;
-    private readonly int _fontWidth;
-    private readonly int _fontHeight;
-    private string _lastMessage = "";
-    private IntPtr _lastFrameTexture;
-    private SDL_Rect _lastFrameLoc;
-    private Dictionary<(char, Color, Color), IntPtr> _cachedGlyphs;
-
-    private Dictionary<Color, SDL_Color> _colours;
-
-    public SDLDisplay(string windowTitle, int fontSize) : base()
+    public void GameLoop()
     {
-        FontSize = fontSize;
-        SDL_Init(SDL_INIT_VIDEO);
-        SDL_ttf.TTF_Init();
-        _font = SDL_ttf.TTF_OpenFont("DejaVuSansMono.ttf", fontSize);
-        SDL_ttf.TTF_SizeUTF8(_font, " ", out _fontWidth, out _fontHeight);
-        
-        int width = ScreenWidth * _fontWidth;
-        int height = ScreenHeight * _fontHeight;
-        _window = SDL_CreateWindow(windowTitle, 100, 100, width, height, SDL_WindowFlags.SDL_WINDOW_SHOWN | SDL_WindowFlags.SDL_WINDOW_INPUT_FOCUS);
-        _renderer = SDL_CreateRenderer(_window, -1, SDL_RendererFlags.SDL_RENDERER_PRESENTVSYNC);
+        GameState? gameState = null;          
+        LoopState state = LoopState.Startup;
+        PregameHandler? pregameHandler = null;
 
-        _colours = [];
+        TitleScreen();  
 
-        _lastFrameLoc = new SDL_Rect
+        DateTime lastPollTime = DateTime.Now;
+
+        while (true) 
         {
-            x = 0,
-            y = _fontHeight,
-            h = (ScreenHeight - 1) * _fontHeight,
-            w = ViewWidth * _fontWidth
-        };
-
-        _cachedGlyphs = new();
-    }
-
-    public override Action? GetCommand(GameState gameState)
-    {
-        while (SDL_PollEvent(out var e) != 0) {
-            switch (e.type)
-            {
-                case SDL_EventType.SDL_QUIT:
-                    return new QuitAction();
-                case SDL_EventType.SDL_TEXTINPUT:
-                    char c;
-                    unsafe
-                    {
-                        c = (char)*e.text.text;                    
-                    }
-                    
-                    return KeyToAction(c, gameState);
-                default:
-                    return new NullAction();
-            }        
-        }
-
-        return new NullAction();
-    }
-
-    public override string QueryUser(string prompt)
-    {
-        string answer = "";
-
-        do 
-        {
-            WriteMessage($"{prompt} {answer}");
-            char ch = WaitForInput();            
-            if (ch == 13)
-                return answer;
-            else if (ch == BACKSPACE)
-                answer = answer.Length > 0 ? answer[..^1] : "";
-            else
-                answer += ch;
-        } 
-        while (true);
-    }
-
-    private static char KeysymToChar(SDL_Keysym keysym) 
-    {
-        return keysym.mod == SDL_Keymod.KMOD_LSHIFT || keysym.mod == SDL_Keymod.KMOD_RSHIFT
-            ? char.ToUpper((char)keysym.sym)
-            : (char)keysym.sym;
-    }
-
-    public override char WaitForInput()
-    {        
-        do 
-        {
-            SDL_WaitEvent(out var e);
-            switch (e.type) 
-            {                
-                case SDL_EventType.SDL_KEYDOWN:                    
-                    if (e.key.keysym.sym == SDL_Keycode.SDLK_LSHIFT || e.key.keysym.sym == SDL_Keycode.SDLK_RSHIFT)
-                        continue;
-                    var ch = e.key.keysym;                                        
-                    SDL_FlushEvent(SDL_EventType.SDL_TEXTINPUT);
-                    return KeysymToChar(ch);                    
-            }
-        } 
-        while (true);        
-    }
-
-    private SDL_Color ToSDLColour(Color colour) 
-    {
-        if (!_colours.TryGetValue(colour, out SDL_Color value)) 
-        {
-            value = new SDL_Color() { 
-                    a = (byte) colour.A, 
-                    r = (byte) colour.R,
-                    g = (byte) colour.G,
-                    b = (byte) colour.B
-            };
-            _colours.Add(colour, value);
-        }
-
-        return value;
-    }
-
-    private void WriteLine(string message, int lineNum, int col, int width)
-    {
-        message = message.PadRight(width);
-        var fontPtr = _font;
-        var fh = _fontHeight;
-        var surface =  SDL_ttf.TTF_RenderText_Shaded(fontPtr, message, ToSDLColour(WHITE), ToSDLColour(BLACK));        
-        var s = (SDL_Surface)Marshal.PtrToStructure(surface, typeof(SDL_Surface))!;
-        
-        var texture = SDL_CreateTextureFromSurface(_renderer, surface);
-        var loc = new SDL_Rect
-        {
-            x = 2 + col * _fontWidth,
-            y = lineNum * fh,
-            h = fh,
-            w = s.w
-        };
-        
-        SDL_FreeSurface(surface);
-        SDL_SetRenderDrawColor(_renderer, 0, 0, 0, 255);
-        SDL_RenderCopy(_renderer, texture, IntPtr.Zero, ref loc);
-    }
-
-    public override void WriteLongMessage(List<string> message)
-    {
-        SDL_RenderClear(_renderer);
-        for (int j = 0; j < message.Count; j++)
-        {
-            WriteLine(message[j], j, 0, ScreenWidth);
-        }
-        SDL_RenderPresent(_renderer);
-        WaitForInput();
-    }
-
-    public override void WriteMessage(string message)
-    {        
-        _lastMessage = message;
-        DrawFrame();
-    }
-
-    private void SDLPut(int row, int col, char ch, Color color) 
-    {
-        var key = (ch, color, BLACK);
-
-        if (!_cachedGlyphs.TryGetValue(key, out IntPtr texture))
-        {
-            var surface =  SDL_ttf.TTF_RenderUNICODE_Shaded(_font, ch.ToString(), ToSDLColour(color), ToSDLColour(BLACK));        
-            var toCache = SDL_CreateTextureFromSurface(_renderer, surface);            
-            SDL_FreeSurface(surface);
-            texture = toCache;
-            _cachedGlyphs.Add(key, texture);
-        }
-
-        var loc = new SDL_Rect { x = col * _fontWidth + 2, y = row * _fontHeight, h = _fontHeight, w = _fontWidth };
-
-        SDL_RenderCopy(_renderer, texture, IntPtr.Zero, ref loc);
-    }
-
-    private IntPtr CreateMainTexture(int playerRow, int playerCol, GameState gameState)
-    {
-        var tw = ViewWidth * _fontWidth;
-        var th = (ScreenHeight - 1) * _fontHeight;
-        var targetTexture = SDL_CreateTexture(_renderer, SDL_PIXELFORMAT_RGBX8888, (int) SDL_TextureAccess.SDL_TEXTUREACCESS_TARGET , tw, th);
-
-        SDL_SetRenderTarget(_renderer, targetTexture);
-
-        int rowOffset = playerRow - PlayerScreenRow;
-        int colOffset = playerCol - PlayerScreenCol;
-        for (int row = 0; row < ScreenHeight - 1; row++)
-        {
-            for (int col = 0; col < ViewWidth; col++)
-            {
-                int vr = row + rowOffset;
-                int vc = col + colOffset;
-                Color color;
-                char ch;
-                if (gameState.Visible!.Contains((gameState.CurrLevel, vr, vc)))                
-                    (color, ch) = TileToGlyph(gameState.Map!.TileAt(vr, vc), true);                
-                else if (gameState.Remebered!.Contains((gameState.CurrLevel, vr, vc)))
-                    (color, ch) = TileToGlyph(gameState.Map!.TileAt(vr, vc), false);
-                else
-                    (color, ch) = (BLACK, ' ');
-                            
-                SDLPut(row, col, ch, color);
-            }
-        }
-        
-        SDLPut(PlayerScreenRow, PlayerScreenCol, '@', WHITE);
-    
-        SDL_SetRenderTarget(_renderer, IntPtr.Zero);
-        
-        return targetTexture;
-    }
-
-    void WriteSideBar(Player player)
-    {
-        var width = ScreenWidth - ViewWidth;
-        WriteLine($"| {player.Name}".PadRight(width), 1, ViewWidth, width);
-        WriteLine($"| HP: {player.CurrHP} ({player.MaxHP})".PadRight(width), 2, ViewWidth, width);
-        
-        string blank = "|".PadRight(ViewWidth);
-        for (int row = 3; row < ScreenHeight; row++)
-        {
-            WriteLine(blank, row, ViewWidth, width);
-        }
-    }
-
-    private void DrawFrame()
-    {
-        SDL_RenderClear(_renderer);
-        WriteLine(_lastMessage, 0, 0, ScreenWidth);
-        if (Player is not null) 
-        {
-            WriteSideBar(Player);
-            SDL_RenderCopy(_renderer, _lastFrameTexture, IntPtr.Zero, ref _lastFrameLoc);
-        }
-        SDL_RenderPresent(_renderer);
-    }
-
-    public override void UpdateDisplay(GameState gameState)
-    {
-        if (Player is null)
-            throw new Exception("Hmm this shouldn't happen");
-
-        _lastFrameTexture = CreateMainTexture(Player.Row, Player.Col, gameState);
-        DrawFrame();        
-    }
-}
-
-internal class BLDisplay : Display, IDisposable
-{        
-    private readonly Dictionary<int, char> KeyToChar = [];
-
-    public BLDisplay(string windowTitle, int fontSize) : base()
-    {
-        FontSize = fontSize;
-        SetUpKeyToCharMap();
-        Terminal.Open();
-        Terminal.Set($"window: size={ScreenWidth}x{ScreenHeight}, title={windowTitle}; font: DejaVuSansMono.ttf, size={FontSize}");            
-    }
-
-    private void SetUpKeyToCharMap()
-    {
-        int curr = (int)TKCodes.InputEvents.TK_A;
-        for (int ch = 'a'; ch <= 'z'; ch++)
-        {
-            KeyToChar.Add(curr++, (char)ch);
-        }
-        curr = (int)TKCodes.InputEvents.TK_1;
-        for (int ch = '1'; ch <= '9'; ch++)
-        {
-            KeyToChar.Add(curr++, (char)ch);
-        }
-        KeyToChar.Add(curr, '0');
-        KeyToChar.Add((int)TKCodes.InputEvents.TK_RETURN_or_ENTER, '\n');
-        KeyToChar.Add((int)TKCodes.InputEvents.TK_SPACE, ' ');
-        KeyToChar.Add((int)TKCodes.InputEvents.TK_BACKSPACE, (char)BACKSPACE);
-        KeyToChar.Add((int)TKCodes.InputEvents.TK_COMMA, ',');
-        KeyToChar.Add((int)TKCodes.InputEvents.TK_PERIOD, '.');
-    }
-
-    public override Action? GetCommand(GameState gameState)
-    {
-        if (Terminal.HasInput())
-        {
-            var ch = WaitForInput();
-            return KeyToAction(ch, gameState);
-        }
-        else
-            return new NullAction();
-    }
-
-    void WriteSideBar()
-    {
-        Terminal.Print(ViewWidth, 1, $"| {Player.Name}".PadRight(ViewWidth));
-        Terminal.Print(ViewWidth, 2, $"| HP: {Player.CurrHP} ({Player.MaxHP})".PadRight(ViewWidth));
-
-        string blank = "|".PadRight(ViewWidth);
-        for (int row = 3; row < ScreenHeight; row++)
-        {
-            Terminal.Print(ViewWidth, row, blank);
-        }
-    }
-
-    public override void UpdateDisplay(GameState gameState)
-    {
-        int rowOffset = Player!.Row - PlayerScreenRow;
-        int colOffset = Player!.Col - PlayerScreenCol;
-        for (int row = 0; row < ScreenHeight - 1; row++)
-        {
-            for (int col = 0; col < ViewWidth; col++)
-            {
-                int vr = row + rowOffset;
-                int vc = col + colOffset;
-
-                if (gameState.Visible!.Contains((gameState.CurrLevel, vr, vc)))
-                {
-                    var (color, ch) = TileToGlyph(gameState.Map!.TileAt(vr, vc), true);
-                    Terminal.Color(color);
-                    Terminal.Put(col, row + 1, ch);
-                }
-                else if (gameState.Remebered!.Contains((gameState.CurrLevel, vr, vc)))
-                {
-                    var (color, ch) = TileToGlyph(gameState.Map!.TileAt(vr, vc), false);
-                    Terminal.Color(color);
-                    Terminal.Put(col, row + 1, ch);
-                }
-                else
-                {
-                    Terminal.Put(col, row + 1, ' ');
-                }
-            }
-        }
-
-        Terminal.Color(WHITE);
-        Terminal.Put(PlayerScreenCol, PlayerScreenRow + 1, '@');
-
-        WriteSideBar();
-
-        Terminal.Refresh();
-    }
-
-    public override void WriteLongMessage(List<string> message)
-    {
-        Terminal.Clear();
-
-        for (int row = 0; row < message.Count; row++)
-        {
-            Terminal.Print(0, row, message[row]);
-        }
-
-        Terminal.Refresh();
-        WaitForInput();
-    }
-
-    public override void WriteMessage(string message)
-    {
-        Terminal.Print(0, 0, message.PadRight(ScreenWidth));
-        Terminal.Refresh();
-    }
-
-    public override string QueryUser(string prompt)
-    {            
-        string answer = "";
-        do
-        {
-            string message = $"{prompt} {answer}";
-            WriteMessage(message);
-
-            var ch = WaitForInput();
-            if (ch == '\n')
-            {
+            var e = PollForEvent();
+            if (e.Type == UIEventType.Quiting)
                 break;
-            }
-            else if (ch == BACKSPACE && answer.Length > 0)
-            {
-                answer = answer[..^1];
-            }
-            else if (ch != '\0')
-            {
-                answer += ch;
-            }
-        }
-        while (true);
 
-        return answer;
-    }
-
-    public override char WaitForInput()
-    {
-        do 
-        {
-            int key = Terminal.Read();
-            
-            if (key == (int)TKCodes.InputEvents.TK_CLOSE)
-                throw new GameQuitException();
-
-            if (KeyToChar.TryGetValue(key, out char value))
+            switch (state)
             {
-                // I feel like there has to be a better way to handle shifted characters
-                // in Bearlib but I haven't found it yet...
-                if (Terminal.Check((int)TKCodes.InputEvents.TK_SHIFT))
-                {
-                    return value switch
+                case LoopState.Startup:
+                    if (e.Type == UIEventType.KeyInput)
                     {
-                        ',' => '<',
-                        '.' => '>',
-                        _ => char.ToUpper(value)
-                    };
-                }
-                else
-                {
-                    return value;
-                }                    
+                        state = LoopState.Pregame;
+                        pregameHandler = new PregameHandler(this);
+                        _longMessage = null;
+                    }
+                    break;
+                case LoopState.Pregame:
+                    if (e.Type == UIEventType.KeyInput && pregameHandler.HandleInput(e))
+                    {
+                        var campaign = pregameHandler.Campaign;
+                        gameState = new GameState()
+                        {
+                            Map = campaign!.Dungeons[0].LevelMaps[campaign.CurrentLevel],
+                            Options = null,
+                            Player = Player,
+                            Campaign = campaign,
+                            CurrLevel = campaign.CurrentLevel,
+                            CurrDungeon = campaign.CurrentDungeon
+                        };
+                        state = LoopState.Playing;
+                    }
+                    break;                
             }
-        }
-        while (true);
-    }
 
-    public override void TitleScreen()
+            if (gameState is not null)
+            {
+                // Maybe move this into gamestate?
+                // or eventually the user input handler class?
+                var c = gameState.Campaign;
+                int currLevel = gameState.CurrLevel;
+                var dungeon = c.Dungeons[gameState.CurrDungeon];
+                var map = dungeon.LevelMaps[currLevel];            
+                var vs = FieldOfView.CalcVisible(Player, map, currLevel);
+                var toShow = vs.Select(v => (v.Item2, v.Item3)).ToHashSet();        
+                dungeon.RememberedSqs = dungeon.RememberedSqs.Union(vs).ToHashSet();
+                dungeon.RememberedSqs = dungeon.RememberedSqs.Union(vs).ToHashSet();
+                gameState.Visible = vs;
+                gameState.Remebered = dungeon.RememberedSqs;
+            }
+            UpdateDisplay(gameState);
+
+            // UpdateView will query for what the user can see
+            // UpdateView(player, gameState);
+            
+            // main loop state machine:
+
+            // oh I need a pre-game routine because I'll ask for 
+            // the player's name etc before the game begins :O
+            // GameState will have methods for loading the game
+            // or starting a new one?
+
+            // 1) waiting for input
+            //    on a key input event I'll want to pass the result 
+            //    to the appropriate handler
+
+            // 2) if idle, then the key input is a player's command
+            //    or are they the same thing and the delegate changes?
+
+            // event can be quit or quit-and-save
+
+            var dd = DateTime.Now - lastPollTime;
+            if (dd.TotalSeconds > 5) 
+            {
+                Console.WriteLine("hello, world?");
+                lastPollTime = DateTime.Now;
+            }
+
+            Thread.Sleep(50);
+        }        
+    }
+}
+
+internal class PregameHandler
+{
+    protected const int BACKSPACE = 8;
+    private const string _prompt = "Who are you?"; 
+    private UserInterface _ui { get; set; }
+    private string _playerName { get; set; } = "";
+    public Campaign? Campaign { get; set; }
+
+    public PregameHandler(UserInterface ui)
     {
-        base.TitleScreen();
-        Terminal.Clear();
-        Terminal.Refresh();
+        _ui = ui;
+        ui.WriteMessage(_prompt);
     }
 
-    public void Dispose()
-    {            
-        Dispose(true);
-    }
-
-    protected virtual void Dispose(bool disposing)
+    (Campaign, int, int) BeginCampaign(Random rng)
     {
-        if (disposing)
+        var dm = new DungeonMaker(rng);
+        var campaign = new Campaign();
+        var wilderness = new Dungeon(0, "You draw a deep breath of fresh air.");
+        var wildernessGenerator = new Wilderness(rng);
+        var map = wildernessGenerator.DrawLevel(257);
+        wilderness.AddMap(map);
+        campaign.AddDungeon(wilderness);
+
+        var mainDungeon = new Dungeon(1, "Musty smells. A distant clang. Danger.");
+        var firstLevel = dm.DrawLevel(100, 40);
+        mainDungeon.AddMap(firstLevel);
+        campaign.AddDungeon(mainDungeon);
+
+        // Find an open floor in the first level of the dungeon
+        // and create a Portal to it in the wilderness
+        var stairs = firstLevel.RandomTile(TileType.Floor, rng);
+        var entrance = map.RandomTile(TileType.Tree, rng);
+        var portal = new Portal("You stand before a looming portal.")
         {
-            Terminal.Close();
-        }            
+            Destination = (1, 0, stairs.Item1, stairs.Item2)
+        };
+        map.SetTile(entrance, portal);
+
+        var exitStairs = new Upstairs("")
+        {
+            Destination = (0, 0, entrance.Item1, entrance.Item2)
+        };
+        firstLevel.SetTile(stairs, exitStairs);
+
+        campaign.CurrentDungeon = 0;
+        campaign.CurrentLevel = 0;
+        return (campaign, entrance.Item1, entrance.Item2);
+    }
+
+    private void SetupGame(string playerName)
+    {
+        var (c, startRow, startCol) = BeginCampaign(new Random());
+        Player player = new Player(playerName, startRow, startCol);          
+        _ui.Player = player;
+        Campaign = c;
+    }
+
+    // Eventually this is going to need state because
+    // we'll be handling picking a character class, etc
+    // for a new game
+    public bool HandleInput(UIEvent e)
+    {
+        if (e.Value == '\n' || e.Value == 13) 
+        {
+            SetupGame(_playerName);
+            _ui.WriteMessage($"Welcome, {_playerName}");
+            return true;
+        }
+        else if (e.Value == BACKSPACE)
+        {
+            _playerName = _playerName.Length > 0 
+                            ? _playerName[..^1] 
+                            : "";
+        }
+        else 
+        {
+            _playerName += e.Value;            
+        }
+
+        _ui.WriteMessage($"{_prompt} {_playerName}");
+        return false;
     }
 }
